cloud_function/requirements.txt
# SmartStudy — Cloud Function dependencies
functions-framework==3.*
google-cloud-storage==2.*
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-vertexai>=2.0.0
pymongo[srv]>=4.6.0
pypdf>=4.0.0
